Add test_handle_soft_bundle_certificates
