+++ crates/sui-core/src/unit_tests/authority_tests.rs
@@ -42,7 +42,6 @@ use sui_types::execution::SharedInput;
 use sui_types::execution_status::{ExecutionFailureStatus, ExecutionStatus};
 use sui_types::gas_coin::GasCoin;
 use sui_types::messages_consensus::ConsensusDeterminedVersionAssignments;
-use sui_types::messages_grpc::HandleSoftBundleCertificatesRequestV3;
 use sui_types::object::Data;
 use sui_types::programmable_transaction_builder::ProgrammableTransactionBuilder;
 use sui_types::randomness_state::get_randomness_state_obj_initial_shared_version;
@@ -4159,7 +4158,7 @@ pub async fn init_state_with_ids_and_object_basics<
     publish_object_basics(state).await
 }
 
-async fn publish_object_basics(state: Arc<AuthorityState>) -> (Arc<AuthorityState>, ObjectRef) {
+pub async fn publish_object_basics(state: Arc<AuthorityState>) -> (Arc<AuthorityState>, ObjectRef) {
     use sui_move_build::BuildConfig;
 
     // add object_basics package object to genesis, since lots of test use it
@@ -5730,7 +5729,7 @@ async fn test_publish_not_a_package_dependency() {
     )
 }
 
-fn create_gas_objects(num: u32, owner: SuiAddress) -> Vec<Object> {
+pub fn create_gas_objects(num: u32, owner: SuiAddress) -> Vec<Object> {
     let mut objects = vec![];
     for _ in 0..num {
         let gas_object_id = ObjectID::random();
@@ -6161,387 +6160,4 @@ async fn test_single_authority_reconfigure() {
     assert_eq!(state.epoch_store_for_testing().epoch(), 0);
     state.reconfigure_for_testing().await;
     assert_eq!(state.epoch_store_for_testing().epoch(), 1);
-}
-
-#[tokio::test]
-async fn test_soft_bundle_failure_modes() {
-    let mut senders = Vec::new();
-    let mut gas_objects = Vec::new();
-    let mut owned_objects = Vec::new();
-    for _i in 0..10 {
-        let (sender, keypair): (_, AccountKeyPair) = get_key_pair();
-        let mut objects = create_gas_objects(2, sender);
-        senders.push((sender, keypair));
-        gas_objects.push(objects.remove(0));
-        owned_objects.push(objects.remove(0));
-    }
-
-    let mut protocol_config =
-        ProtocolConfig::get_for_version(ProtocolVersion::max(), Chain::Unknown);
-    protocol_config.set_enable_soft_bundle(true);
-    protocol_config.set_max_soft_bundle_size(3);
-    let authority = TestAuthorityBuilder::new()
-        .with_reference_gas_price(1000)
-        .with_protocol_config(protocol_config)
-        .build()
-        .await;
-
-    authority.insert_genesis_objects(&gas_objects).await;
-    authority.insert_genesis_objects(&owned_objects).await;
-
-    let (authority, package) = publish_object_basics(authority).await;
-
-    let shared_object = {
-        let effects = call_move_(
-            &authority,
-            None,
-            &gas_objects[3].id(),
-            &senders[3].0,
-            &senders[3].1,
-            &package.0,
-            "object_basics",
-            "share",
-            vec![],
-            vec![],
-            true,
-        )
-        .await
-        .unwrap();
-        effects.status().unwrap();
-        let shared_object_id = effects.created()[0].0 .0;
-        authority
-            .get_object(&shared_object_id)
-            .await
-            .unwrap()
-            .unwrap()
-    };
-    let initial_shared_version = shared_object.version();
-
-    // Create a single validator cluster.
-    let consensus_address = "/ip4/127.0.0.1/tcp/0/http".parse().unwrap();
-    let server = AuthorityServer::new_for_test(
-        "/ip4/127.0.0.1/tcp/0/http".parse().unwrap(),
-        authority,
-        consensus_address,
-    );
-    let authority = server.state.clone();
-    let _metrics = server.metrics.clone();
-    let server_handle = server.spawn_for_test().await.unwrap();
-    let client = NetworkAuthorityClient::connect(server_handle.address())
-        .await
-        .unwrap();
-
-    let signed_tx_into_certificate = |transaction: Transaction| async {
-        let epoch_store = authority.load_epoch_store_one_call_per_task();
-        let committee = authority.clone_committee_for_testing();
-        let mut sigs = vec![];
-
-        let transaction = epoch_store.verify_transaction(transaction).unwrap();
-        let response = authority
-            .handle_transaction(&epoch_store, transaction.clone())
-            .await
-            .unwrap();
-        let vote = response.status.into_signed_for_testing();
-        sigs.push(vote.clone());
-        if let Ok(cert) =
-            CertifiedTransaction::new(transaction.clone().into_message(), sigs.clone(), &committee)
-        {
-            return cert
-                .try_into_verified_for_testing(&committee, &Default::default())
-                .unwrap();
-        }
-        panic!("Failed to create certificate");
-    };
-
-    let rgp = authority.reference_gas_price_for_testing().unwrap();
-
-    // Case 0: submit an empty soft bundle.
-    {
-        let response = client
-            .handle_soft_bundle_certificates_v3(
-                HandleSoftBundleCertificatesRequestV3 {
-                    certificates: vec![],
-                    wait_for_effects: true,
-                    include_events: false,
-                    include_auxiliary_data: false,
-                    include_input_objects: false,
-                    include_output_objects: false,
-                },
-                None,
-            )
-            .await;
-        assert!(response.is_err());
-        assert!(match response.unwrap_err() {
-            SuiError::NoCertificateProvidedError { .. } => true,
-            _ => false,
-        })
-    }
-
-    // Case 1: submit a soft bundle with more txs than the limit.
-    // The bundle should be rejected.
-    {
-        let mut certificates: Vec<CertifiedTransaction> = vec![];
-        for i in 0..5 {
-            let owned_object_ref = authority
-                .get_object(&owned_objects[i].id())
-                .await
-                .unwrap()
-                .unwrap()
-                .compute_object_reference();
-            let gas_object_ref = authority
-                .get_object(&gas_objects[i].id())
-                .await
-                .unwrap()
-                .unwrap()
-                .compute_object_reference();
-            let data = TransactionData::new_transfer(
-                senders[i + 1].0,
-                owned_object_ref,
-                senders[i].0,
-                gas_object_ref,
-                rgp * TEST_ONLY_GAS_UNIT_FOR_TRANSFER,
-                rgp,
-            );
-            let signed = to_sender_signed_transaction(data, &senders[i].1);
-            certificates.push(signed_tx_into_certificate(signed).await.into());
-        }
-        let response = client
-            .handle_soft_bundle_certificates_v3(
-                HandleSoftBundleCertificatesRequestV3 {
-                    certificates,
-                    wait_for_effects: true,
-                    include_events: false,
-                    include_auxiliary_data: false,
-                    include_input_objects: false,
-                    include_output_objects: false,
-                },
-                None,
-            )
-            .await;
-        assert!(response.is_err());
-        assert!(match response.unwrap_err() {
-            SuiError::UserInputError {
-                error: UserInputError::TooManyTransactionsInSoftBundle { .. },
-            } => true,
-            _ => false,
-        })
-    }
-
-    // Case 2: submit a soft bundle with tx containing no shared object.
-    // The bundle should be rejected.
-    {
-        let owned_object_ref = authority
-            .get_object(&owned_objects[5].id())
-            .await
-            .unwrap()
-            .unwrap()
-            .compute_object_reference();
-        let gas_object_ref = authority
-            .get_object(&gas_objects[5].id())
-            .await
-            .unwrap()
-            .unwrap()
-            .compute_object_reference();
-        let data = TransactionData::new_transfer(
-            senders[6].0,
-            owned_object_ref,
-            senders[5].0,
-            gas_object_ref,
-            rgp * TEST_ONLY_GAS_UNIT_FOR_TRANSFER,
-            rgp,
-        );
-        let signed = to_sender_signed_transaction(data, &senders[5].1);
-        let response = client
-            .handle_soft_bundle_certificates_v3(
-                HandleSoftBundleCertificatesRequestV3 {
-                    certificates: vec![signed_tx_into_certificate(signed).await.into()],
-                    wait_for_effects: true,
-                    include_events: false,
-                    include_auxiliary_data: false,
-                    include_input_objects: false,
-                    include_output_objects: false,
-                },
-                None,
-            )
-            .await;
-        assert!(response.is_err());
-        assert!(match response.unwrap_err() {
-            SuiError::UserInputError {
-                error: UserInputError::NoSharedObjectError { .. },
-            } => true,
-            _ => false,
-        })
-    }
-
-    // Case 3: submit a soft bundle with txs of different gas prices.
-    // The bundle should be rejected.
-    {
-        let cert0 = {
-            let gas_object_ref = authority
-                .get_object(&gas_objects[6].id())
-                .await
-                .unwrap()
-                .unwrap()
-                .compute_object_reference();
-            let data = TransactionData::new_move_call(
-                senders[6].0,
-                package.0,
-                ident_str!("object_basics").to_owned(),
-                ident_str!("set_value").to_owned(),
-                /* type_args */ vec![],
-                gas_object_ref,
-                /* args */
-                vec![
-                    CallArg::Object(ObjectArg::SharedObject {
-                        id: shared_object.id(),
-                        initial_shared_version,
-                        mutable: true,
-                    }),
-                    CallArg::Pure(11u64.to_le_bytes().to_vec()),
-                ],
-                TEST_ONLY_GAS_UNIT_FOR_OBJECT_BASICS * rgp,
-                rgp,
-            )
-            .unwrap();
-            let signed = to_sender_signed_transaction(data, &senders[6].1);
-            signed_tx_into_certificate(signed).await
-        };
-        let cert1 = {
-            let gas_object_ref = authority
-                .get_object(&gas_objects[7].id())
-                .await
-                .unwrap()
-                .unwrap()
-                .compute_object_reference();
-            let data = TransactionData::new_move_call(
-                senders[7].0,
-                package.0,
-                ident_str!("object_basics").to_owned(),
-                ident_str!("set_value").to_owned(),
-                /* type_args */ vec![],
-                gas_object_ref,
-                /* args */
-                vec![
-                    CallArg::Object(ObjectArg::SharedObject {
-                        id: shared_object.id(),
-                        initial_shared_version,
-                        mutable: true,
-                    }),
-                    CallArg::Pure(12u64.to_le_bytes().to_vec()),
-                ],
-                TEST_ONLY_GAS_UNIT_FOR_OBJECT_BASICS * rgp,
-                rgp + 1,
-            )
-            .unwrap();
-            let signed = to_sender_signed_transaction(data, &senders[7].1);
-            signed_tx_into_certificate(signed).await
-        };
-        let response = client
-            .handle_soft_bundle_certificates_v3(
-                HandleSoftBundleCertificatesRequestV3 {
-                    certificates: vec![cert0.into(), cert1.into()],
-                    wait_for_effects: true,
-                    include_events: false,
-                    include_auxiliary_data: false,
-                    include_input_objects: false,
-                    include_output_objects: false,
-                },
-                None,
-            )
-            .await;
-        assert!(response.is_err());
-        assert!(match response.unwrap_err() {
-            SuiError::UserInputError {
-                error: UserInputError::GasPriceMismatchError { .. },
-            } => true,
-            _ => false,
-        })
-    }
-
-    // Case 4: submit a soft bundle with txs whose consensus message has been processed.
-    // The bundle should be rejected.
-    {
-        let cert0 = {
-            let gas_object_ref = authority
-                .get_object(&gas_objects[8].id())
-                .await
-                .unwrap()
-                .unwrap()
-                .compute_object_reference();
-            let data = TransactionData::new_move_call(
-                senders[8].0,
-                package.0,
-                ident_str!("object_basics").to_owned(),
-                ident_str!("set_value").to_owned(),
-                /* type_args */ vec![],
-                gas_object_ref,
-                /* args */
-                vec![
-                    CallArg::Object(ObjectArg::SharedObject {
-                        id: shared_object.id(),
-                        initial_shared_version,
-                        mutable: true,
-                    }),
-                    CallArg::Pure(11u64.to_le_bytes().to_vec()),
-                ],
-                TEST_ONLY_GAS_UNIT_FOR_OBJECT_BASICS * rgp,
-                rgp,
-            )
-            .unwrap();
-            let signed = to_sender_signed_transaction(data, &senders[8].1);
-            signed_tx_into_certificate(signed).await
-        };
-        let cert1 = {
-            let gas_object_ref = authority
-                .get_object(&gas_objects[9].id())
-                .await
-                .unwrap()
-                .unwrap()
-                .compute_object_reference();
-            let data = TransactionData::new_move_call(
-                senders[9].0,
-                package.0,
-                ident_str!("object_basics").to_owned(),
-                ident_str!("set_value").to_owned(),
-                /* type_args */ vec![],
-                gas_object_ref,
-                /* args */
-                vec![
-                    CallArg::Object(ObjectArg::SharedObject {
-                        id: shared_object.id(),
-                        initial_shared_version,
-                        mutable: true,
-                    }),
-                    CallArg::Pure(12u64.to_le_bytes().to_vec()),
-                ],
-                TEST_ONLY_GAS_UNIT_FOR_OBJECT_BASICS * rgp,
-                rgp,
-            )
-            .unwrap();
-            let signed = to_sender_signed_transaction(data, &senders[9].1);
-            signed_tx_into_certificate(signed).await
-        };
-        send_batch_consensus_no_execution(&authority, &vec![cert0.clone(), cert1.clone()], true)
-            .await;
-        let response = client
-            .handle_soft_bundle_certificates_v3(
-                HandleSoftBundleCertificatesRequestV3 {
-                    certificates: vec![cert0.into(), cert1.into()],
-                    wait_for_effects: true,
-                    include_events: false,
-                    include_auxiliary_data: false,
-                    include_input_objects: false,
-                    include_output_objects: false,
-                },
-                None,
-            )
-            .await;
-        assert!(response.is_err());
-        assert!(match response.unwrap_err() {
-            SuiError::UserInputError {
-                error: UserInputError::CeritificateAlreadyProcessed { .. },
-            } => true,
-            _ => false,
-        })
-    }
 }
